refactor(fragments): modularize sync flow and enforce legacy importer
